+++ sdk/messaging/azeventhubs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release History
 
+## 1.3.1 (TBD)
+
+### Features Added
+
+- ProducerClient and ConsumerClient can use endpoint formats that the Azure CLI or ARM return, like https://<your-namespace>.servicebus.windows.net:443. (PR#TBD)
+
 ## 1.3.0 (2025-02-11)
 
 ### Features Added